CassandraAdapter

+++ anyline-data-jdbc-dialect/anyline-data-jdbc-cassandra/src/main/java/org/anyline/data/jdbc/cassandra/CassandraAdapter.java
@@ -50,7 +50,7 @@ public class CassandraAdapter extends SQLAdapter implements JDBCAdapter, Initial
         }
         PageNavi navi = run.getPageNavi();
         if(null != navi){
-            int limit = navi.getLastRow() - navi.getFirstRow() + 1;
+            long limit = navi.getLastRow() - navi.getFirstRow() + 1;
             if(limit < 0){
                 limit = 0;
             }
